// src/shared/types/character/AttributesForm.ts
export type AttributesForm = {
  CON: number;
  FR: number;
  DEX: number;
  AGI: number;
  INT: number;
  WILL: number;
  PER: number;
  CAR: number;
};


export type AttributesModForm = {
  con_mod: number;
  fr_mod: number;
  dex_mod: number;
  agi_mod: number;
  int_mod: number;
  will_mod: number;
  per_mod: number;
  car_mod: number;
};